Updating tests : test_coingecko_crypto
